:ok_hand: [#81] PR feedback
